requirements.txt
greenlet==3.1.1
playwright==1.48.0
pyee==12.0.0
python-dotenv==1.0.1
typing_extensions==4.12.2
